split load logic from resourceManager into a new LoadOperation class
