+++ javascript/workflow/src/resources/DataPackage/ResourceManager.ts
@@ -2,18 +2,14 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import type {
-	DataPackageSchema,
-	Profile,
-	ResourceSchema,
-} from '@datashaper/schema'
+import type { DataPackageSchema, Profile } from '@datashaper/schema'
 import { KnownProfile } from '@datashaper/schema'
 import type { Observable } from 'rxjs'
 import { BehaviorSubject, map } from 'rxjs'
 
 import type { Resource } from '../Resource.js'
-import { ResourceReference } from '../ResourceReference.js'
 import type { ProfileHandler } from '../types.js'
+import { LoadResourcesOperation } from './LoadResourcesOperation.js'
 import { CodebookProfile } from './profiles/CodebookProfile.js'
 import { DataTableProfile } from './profiles/DataTableProfile.js'
 import { TableBundleProfile } from './profiles/TableBundleProfile.js'
@@ -43,6 +39,10 @@ export class ResourceManager {
 		[KnownProfile.Codebook, new CodebookProfile()],
 		[KnownProfile.Workflow, new WorkflowProfile()],
 	] as [Profile, ProfileHandler][])
+
+	public get profileHandlers() {
+		return this._profileHandlers
+	}
 
 	/**
 	 * The top-level resources observable
@@ -256,251 +256,28 @@ export class ResourceManager {
 	 * @returns The loaded datapackage schema
 	 */
 	public async load(files: Map<string, Blob>): Promise<DataPackageSchema> {
-		const resolveResource = (
-			entry: string | ResourceSchema,
-		): Resource | undefined => {
-			const isRef = this.isReference(entry)
-			// if entry is a string, it's likely a path, otherwise check for a name
-			let key = typeof entry === 'string' ? entry : entry.name
-			if (isRef) {
-				key = entry.path
-			}
-			const resource = this.getResource(key) || this.getResourceByPath(key)
-
-			if (isRef) {
-				const reference = new ResourceReference()
-				reference.target = resource
-				reference.rel = entry.rel
-				return reference
-			} else {
-				return resource
-			}
-		}
-
 		this.clear()
 		this._files = files
 
-		const dataPackage = await this.getDataPackageSchema()
-		/**
-		 * Step 1: Read all Schemas
-		 */
-		const schemas: { schema: ResourceSchema; path: string }[] =
-			await this.readSchemas(dataPackage)
+		const [dataPackage, resources, topLevelResources] =
+			await new LoadResourcesOperation(files, this).execute()
 
-		/**
-		 * Step 2: Reify Instances
-		 */
-		const nonReferenceSchemas = schemas.filter(s => !this.isReference(s.schema))
-		const resources = await this.reifySchemas(nonReferenceSchemas)
-
-		/**
-		 * Step 3: Track resources by name and path
-		 */
-		for (const { resource, path } of resources) {
-			this.registerResource(resource)
-			this._resourceByPath.set(path, resource)
-		}
-
-		/**
-		 * Step 4: Link Resources and Resolve References
-		 */
-		for (const { resource, schema } of resources) {
-			const sources = (schema.sources?.map(resolveResource).filter(t => !!t) ??
-				[]) as Resource[]
-
-			resource.sources = sources
-		}
-
-		const topLevelResources = dataPackage.resources
-			.map(resolveResource)
-			.filter(t => !!t) as Resource[]
-
-		this._resources$.next(resources.map(r => r.resource))
+		this._resources$.next(resources)
 		this._topResources$.next(topLevelResources)
-		resources.forEach(r =>
-			r.resource.onChange(() => this._topResources$.next(this.topResources)),
+		topLevelResources.forEach(r =>
+			r.onChange(() => this._topResources$.next(this.topResources)),
 		)
 		return dataPackage
 	}
 
-	private async readSchemas(
-		dataPackage: DataPackageSchema,
-	): Promise<{ schema: ResourceSchema; path: string }[]> {
-		const nameToPath = new Map<string, string>()
-
-		// Determine the name of a resource, try to not collide if possible
-		const resourceName = (res: ResourceSchema) => {
-			if (res.name) {
-				return res.name
-			} else {
-				const profile = res.profile || 'resource'
-				let candidate = `${profile}.json`
-				let index = 0
-				while (nameToPath.has(candidate)) {
-					candidate = `${profile}-${++index}.json`
-				}
-				return candidate
-			}
-		}
-
-		// Resolve a schema entry to a schema object
-		const resolveSchema = async (
-			entry: string | ResourceSchema,
-		): Promise<ResourceSchema | undefined> => {
-			if (this.isReference(entry)) {
-				entry = entry.path
-			} else if (typeof entry !== 'string') {
-				return entry
-			}
-
-			if (nameToPath.has(entry)) {
-				// this is a duplicate, return undefined
-				return undefined
-			}
-			return this.readResource(entry)
-		}
-
-		const resolvePath = (
-			entry: string | ResourceSchema,
-			schema: ResourceSchema,
-			parentPath: string,
-		) => {
-			if (this.isReference(entry)) {
-				return entry.path
-			}
-			return typeof entry === 'string'
-				? entry
-				: `${parentPath}${resourceName(schema)}`
-		}
-
-		const readEntry = async (
-			entry: string | ResourceSchema,
-			parentPath: string,
-		): Promise<{ schema: ResourceSchema; path: string } | undefined> => {
-			const schema = await resolveSchema(entry)
-			if (!schema) {
-				return undefined
-			}
-			schema.name = schema.name || resourceName(schema)
-			const path = resolvePath(entry, schema, parentPath)
-
-			if (nameToPath.has(schema.name)) {
-				const existingPath = nameToPath.get(schema.name)
-				if (existingPath !== path) {
-					throw new Error(
-						`duplicate resource name ${schema.name} found in two different paths: ${existingPath} and ${path}`,
-					)
-				}
-				return undefined
-			}
-			nameToPath.set(schema.name, path)
-			return { schema, path }
-		}
-
-		const readEntries = async (
-			entries: (ResourceSchema | string)[],
-			parentPath = '',
-		): Promise<Array<{ schema: ResourceSchema; path: string }>> => {
-			const result = await Promise.all(
-				entries.map(e => readEntry(e, parentPath)),
-			)
-			return result.filter(t => !!t) as Array<{
-				schema: ResourceSchema
-				path: string
-			}>
-		}
-
-		// Read the top-level schemas
-		const schemas = await readEntries(dataPackage.resources)
-
-		// Read all children scemas
-		for (let i = 0; i < schemas.length; i++) {
-			const { schema, path } = schemas[i]!
-			const sources = schema.sources ?? []
-			schemas.push(...(await readEntries(sources, `${path}/`)))
-		}
-
-		return schemas
-	}
-
-	private reifySchemas(
-		schemas: Array<{ schema: ResourceSchema; path: string }>,
-	): Promise<
-		Array<{ path: string; resource: Resource; schema: ResourceSchema }>
-	> {
-		return Promise.all(
-			schemas.map(({ schema, path }) => {
-				// override sources so that we can link them together in a second pass
-				return this.constructResource({ ...schema, sources: [] }).then(
-					resource => ({
-						resource,
-						schema,
-						path,
-					}),
-				)
-			}),
-		)
-	}
-
-	private async readResource(entry: string): Promise<ResourceSchema> {
-		// if the item is a string, look up the resource in the files map
-		let result: ResourceSchema | undefined
-
-		const blob = this._files.get(entry)
-		const resourceText = await blob?.text()
-		if (resourceText != null) {
-			try {
-				result = JSON.parse(resourceText) as ResourceSchema
-			} catch (e) {
-				console.error(`error parsing resource ${entry}`, e)
-			}
-		}
-		if (result == null) {
-			throw new Error(`could not resolve resource "${entry}"`)
-		}
-		return result
-	}
-
-	private async constructResource(schema: ResourceSchema): Promise<Resource> {
-		let result: Resource | undefined
-		if (schema.profile == null) {
-			throw new Error('schema has no profile')
-		}
-		const handler = this._profileHandlers.get(schema.profile)
-		if (handler != null) {
-			result = await handler.createInstance(schema, this)
-		}
-		if (result == null) {
-			throw new Error(
-				`could not construct resource with profile "${schema.profile}", are you missing a resource handler?"`,
-			)
-		}
-		return result
-	}
-
-	private registerResource(resource: Resource) {
+	public registerResource(resource: Resource, path?: string | undefined) {
 		if (this._resourceByName.has(resource.name)) {
 			throw new Error(`duplicate resource name: ${resource.name}`)
 		}
 		this._resourceByName.set(resource.name, resource)
-	}
 
-	private async getDataPackageSchema(): Promise<DataPackageSchema> {
-		const dataPackageBlob = this._files.get('datapackage.json')
-		if (dataPackageBlob == null) {
-			throw new Error('file archive must contain datapackage.json')
+		if (path != null) {
+			this._resourceByPath.set(path, resource)
 		}
-		return JSON.parse(await dataPackageBlob.text()) as DataPackageSchema
-	}
-
-	private isReference(
-		entry: ResourceSchema | string,
-	): entry is ResourceSchema & { path: string } {
-		return (
-			typeof entry !== 'string' &&
-			entry.profile == null &&
-			entry.path != null &&
-			typeof entry.path === 'string'
-		)
 	}
 }
