+++ javascript/workflow/src/resources/types.ts
@@ -45,12 +45,12 @@ export interface ProfileHandler {
 		manager: ResourceManager,
 	): Promise<Resource>
 
-	// /**
-	//  * Save custom resources into the files map.
-	//  * @param files - the data files in the package
-	//  * @returns the list of top-level resources to write into datapackage.json::resources
-	//  */
-	// save?: (data: Resource, files: Map<string, Blob>) => Promise<string[]>
+	/**
+	 * Save custom resources into the files map.
+	 * @param files - the data files in the package
+	 * @returns the list of top-level resources to write into datapackage.json::resources
+	 */
+	save?: (data: Resource, files: Map<string, Blob>) => Promise<string[]>
 
 	// /**
 	//  * Load a specific resource
